짐보관 픽업 통신 선언, 리뷰 작성 미완

+++ Travely_Manager/app/src/main/java/com/hyeran/android/travely_manager/network/NetworkService.kt
@@ -50,7 +50,7 @@ interface NetworkService{
     @PUT("/api/owner/reserve/{reserveIdx}")
     fun putStorePickUpResponse(
             @Header("jwt") jwt: String?,
-            @Path("reserveIdx") reserveIdx: Int
+            @Path("reserveIdx") reserveIdx: Long
     ) : Call<Any>
 
     // 예약코드 조회
